created command interface and classes for every command

+++ sitoop1project22621621/src/GrammarCommands.java
@@ -1,10 +1,10 @@
 
 public interface GrammarCommands {
     void list();
-    void print(int id);
-    void save(int id, String fileName);
-    void addRule(int id, String rule);
-    void removeRule(int id, int number) throws Exception;
+    void print(long id);
+    void save(long id, String fileName);
+    void addRule(long id, String rule);
+    void removeRule(long id, int number) throws Exception;
     //void union(int id1, int id2);
     //void concat(int id1, int id2);
    // void chomsky(int id);
